pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pydataform"
version = "0.0.1"
description = "A Python library for simplifying Google Cloud Dataform operations"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "Apache-2.0"}
keywords = ["google-cloud", "dataform", "bigquery", "data-engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "google-cloud-dataform>=0.6.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/pydataform"
Documentation = "https://github.com/yourusername/pydataform#readme"
Repository = "https://github.com/yourusername/pydataform.git"
Issues = "https://github.com/yourusername/pydataform/issues"

[tool.setuptools]
packages = ["pydataform"]

[tool.setuptools_scm]
write_to = "pydataform/_version.py"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 